refactor: rename index.rs to page.rs
